+++ cpp/USRP_server_diagnostic.cpp
@@ -5,14 +5,14 @@ void set_this_thread_name(std::string thread_name){
 
     boost::log::core::get()->add_thread_attribute("ThreadName",
             boost::log::attributes::constant< std::string >(thread_name));
-            
+
 }
 
 //print on screen error description
 void interptet_rx_error(uhd::rx_metadata_t::error_code_t error){
 
     switch(error){
-    
+
         case uhd::rx_metadata_t::error_code_t::ERROR_CODE_NONE:
             print_warning("RX error: Interpreter called on a error free packet.");
             break;
@@ -40,10 +40,10 @@ void interptet_rx_error(uhd::rx_metadata_t::error_code_t error){
 }
 
 int get_rx_errors(uhd::rx_metadata_t *metadata, bool verbose){
-    
+
     //initialize error counter
     int error = 0;
-    
+
     //maximum result will be one in this configuration
     if(metadata->error_code != uhd::rx_metadata_t::error_code_t::ERROR_CODE_NONE){
         error=1;
@@ -54,17 +54,18 @@ int get_rx_errors(uhd::rx_metadata_t *metadata, bool verbose){
     }
     //if(metadata->fragment_offset!=0)print_warning("RX metadata shows fragmentation of the single packet transmission. Try reducing the buffer length.");
     if(metadata->more_fragments)print_error("RX buffer mismatch!");
-    
+
     return error;
 }
 
 int get_tx_error(uhd::async_metadata_t *async_md, bool verbose){
     int error = 0;
     switch(async_md->event_code){
+        /*
         case 0 :
             if(verbose)print_warning("TX async metadata non init");
             break;
-    
+        */
         case uhd::async_metadata_t::EVENT_CODE_TIME_ERROR:
             if(verbose)print_warning("TX thread encountered a time error");
             error++;
@@ -108,10 +109,10 @@ void print_params(usrp_param my_parameter){
     bool B_TXRX = true;
     bool A_RX2 = true;
     bool B_RX2 = true;
-    
+
     std::stringstream ss;
     boost::format position_formatting("%=1.1e");
-    
+
     ss << "\033[40;1;37m                  A N A L O G    P A R A M E T E R S                    \033[0m"<<std::endl;
     ss               <<"------------------------------------------------------------------------"<<std::endl;
     ss               <<"        device on-server ID: "<< my_parameter.usrp_number << "                "<<std::endl;
@@ -125,80 +126,80 @@ void print_params(usrp_param my_parameter){
     ss <<" param "<<std::endl;
 
     ss << "\033[47;1;30m MODE  \033[0m"<<"\t       ";
-    
+
     ss << ant_mode_to_str(my_parameter.A_TXRX.mode)<< "\t\t";
     ss << ant_mode_to_str(my_parameter.A_RX2.mode)<< "\t\t";
     ss << ant_mode_to_str(my_parameter.B_TXRX.mode)<< "\t\t";
     ss << ant_mode_to_str(my_parameter.B_RX2.mode)<< std::endl;
-    
+
     ss << std::scientific;
-    
+
     if(my_parameter.A_TXRX.mode == OFF) A_TXRX = false;
     if(my_parameter.B_TXRX.mode == OFF) B_TXRX = false;
     if(my_parameter.A_RX2.mode == OFF) A_RX2 = false;
     if(my_parameter.B_RX2.mode == OFF) B_RX2 = false;
 
     ss << "\033[47;1;30m RATE  \033[0m"<<"     ";
-    
+
     ss << (A_TXRX?(position_formatting % (my_parameter.A_TXRX.rate/1e6)).str():(std::string)"  -    ") << "          ";
     ss << (A_RX2?(position_formatting % (my_parameter.A_RX2.rate/1e6)).str():(std::string)"  -    ")   << "          ";
     ss << (B_TXRX?(position_formatting % (my_parameter.B_TXRX.rate/1e6)).str():(std::string)"  -    ") << "          ";
     ss << (B_RX2?(position_formatting % (my_parameter.B_RX2.rate/1e6)).str():(std::string)"  -    ") << "\t[MHz]"<<std::endl;
-    
-    
+
+
     ss << "\033[47;1;30m GAIN  \033[0m"<<"     ";
-    
+
     ss << (A_TXRX?(position_formatting % (my_parameter.A_TXRX.gain/1.)).str():(std::string)"  -    ") << "          ";
     ss << (A_RX2?(position_formatting % (my_parameter.A_RX2.gain/1.)).str():(std::string)"  -    ")   << "          ";
     ss << (B_TXRX?(position_formatting % (my_parameter.B_TXRX.gain/1.)).str():(std::string)"  -    ") << "          ";
     ss << (B_RX2?(position_formatting % (my_parameter.B_RX2.gain/1.)).str():(std::string)"  -    ") << "\t[dB]"<<std::endl;
-    
+
     ss << "\033[47;1;30m TONE  \033[0m"<<"     ";
-    
+
     ss << (A_TXRX?(position_formatting % (my_parameter.A_TXRX.tone/1e6)).str():(std::string)"  -    ") << "          ";
     ss << (A_RX2?(position_formatting % (my_parameter.A_RX2.tone/1e6)).str():(std::string)"  -    ")   << "          ";
     ss << (B_TXRX?(position_formatting % (my_parameter.B_TXRX.tone/1e6)).str():(std::string)"  -    ") << "          ";
     ss << (B_RX2?(position_formatting % (my_parameter.B_RX2.tone/1e6)).str():(std::string)"  -    ") << "\t[MHz]"<<std::endl;
-    
+
     ss << "\033[47;1;30m DELAY \033[0m"<<"     ";
-    
+
     ss << (A_TXRX?(position_formatting % (my_parameter.A_TXRX.delay*1e3)).str():(std::string)"  -    ") << "          ";
     ss << (A_RX2?(position_formatting % (my_parameter.A_RX2.delay*1e3)).str():(std::string)"  -    ")   << "          ";
     ss << (B_TXRX?(position_formatting % (my_parameter.B_TXRX.delay*1e3)).str():(std::string)"  -    ") << "          ";
     ss << (B_RX2?(position_formatting % (my_parameter.B_RX2.delay*1e3)).str():(std::string)"  -    ") << "\t[msec]"<<std::endl;
-    
+
     ss << "\033[47;1;30m  BW   \033[0m"<<"     ";
-    
+
     ss << (A_TXRX?(position_formatting % (my_parameter.A_TXRX.bw/1e6)).str():(std::string)"  -    ") << "          ";
     ss << (A_RX2?(position_formatting % (my_parameter.A_RX2.bw/1e6)).str():(std::string)"  -    ")   << "          ";
     ss << (B_TXRX?(position_formatting % (my_parameter.B_TXRX.bw/1e6)).str():(std::string)"  -    ") << "          ";
     ss << (B_RX2?(position_formatting % (my_parameter.B_RX2.bw/1e6)).str():(std::string)"  -    ") << "\t[MHz]"<<std::endl;
-    
+
     ss << "\033[47;1;30m #SIGs \033[0m"<<"     ";
-    
+
     ss << (A_TXRX?(position_formatting % (my_parameter.A_TXRX.wave_type.size()/1.)).str():(std::string)"  -    ") << "          ";
     ss << (A_RX2?(position_formatting % (my_parameter.A_RX2.wave_type.size()/1.)).str():(std::string)"  -    ")   << "          ";
     ss << (B_TXRX?(position_formatting % (my_parameter.B_TXRX.wave_type.size()/1.)).str():(std::string)"  -    ") << "          ";
     ss << (B_RX2?(position_formatting % (my_parameter.B_RX2.wave_type.size()/1.)).str():(std::string)"  -    ") << "\t[#]"<<std::endl;
-    
+
     ss << "\033[47;1;30m #samp \033[0m"<<"     ";
-    
+
     ss << (A_TXRX?(position_formatting % (my_parameter.A_TXRX.samples/1.)).str():(std::string)"  -    ") << "          ";
     ss << (A_RX2?(position_formatting % (my_parameter.A_RX2.samples/1.)).str():(std::string)"  -    ")   << "          ";
     ss << (B_TXRX?(position_formatting % (my_parameter.B_TXRX.samples/1.)).str():(std::string)"  -    ") << "          ";
     ss << (B_RX2?(position_formatting % (my_parameter.B_RX2.samples/1.)).str():(std::string)"  -    ") << "\t[#]"<<std::endl;
-    
+
     ss << "\033[47;1;30m tun_m \033[0m"<<"     ";
     ss << (A_TXRX?(my_parameter.A_TXRX.tuning_mode?"fractio":"integer"):(std::string)"  -    ") << "          ";
     ss << (A_RX2?(my_parameter.A_TXRX.tuning_mode?"fractio":"integer"):(std::string)"  -    ")   << "          ";
     ss << (B_TXRX?(my_parameter.A_TXRX.tuning_mode?"fractio":"integer"):(std::string)"  -    ") << "          ";
     ss << (B_RX2?(my_parameter.A_TXRX.tuning_mode?"fractio":"integer"):(std::string)"  -    ") << "\t[-]"<<std::endl;
-    
+
     ss << std::endl;
     ss << "\033[40;1;37m                  S I G N A L S    P A R A M E T E R S                  \033[0m"<<std::endl<<std::endl;
 
     if(A_TXRX){
-        
+
         ss <<   "\033[40;1;32mRF A TX/RX                                        No. of signals:    ";
         ss << my_parameter.A_TXRX.wave_type.size()<<" \033[0m"<<std::endl<<std::endl;
         ss <<   "\033[40;1;32mRF Buffer lenght:  "<<my_parameter.A_TXRX.buffer_len;
@@ -206,14 +207,14 @@ void print_params(usrp_param my_parameter){
             ss <<                                   "         PFB points:   "<<my_parameter.A_TXRX.fft_tones;
             ss <<                                                             "     PFB average:  "<<  my_parameter.A_TXRX.pf_average<<" \033[0m"<<std::endl<<std::endl;
         }else{
-            ss <<"                                            "<<" \033[0m"<<std::endl<<std::endl;                     
+            ss <<"                                            "<<" \033[0m"<<std::endl<<std::endl;
         }
-        
+
         ss << "\033[47;1;30m   No    \033[0m"<<" "<< "\033[47;1;30m   AMP   \033[0m"<<" "<<"\033[47;1;30m  FREQ   \033[0m"<<" ";
         ss << "\033[47;1;30m  2 FRQ  \033[0m"<<" "<< "\033[47;1;30m  STEPS  \033[0m"<<" "<<"\033[47;1;30m  LAPSE  \033[0m"<< "";
         ss << "\033[47;1;30m   TYPE  \033[0m"<<std::endl;
-        for(int i = 0; i<my_parameter.A_TXRX.wave_type.size();i++){
-        
+        for(size_t i = 0; i<my_parameter.A_TXRX.wave_type.size();i++){
+
             ss << "    "<<(position_formatting % (i)).str() <<"      ";
             ss << (position_formatting % (my_parameter.A_TXRX.ampl[i]/1.)).str() <<"   ";
             ss << (position_formatting % (my_parameter.A_TXRX.freq[i]/1e6)).str() <<"   ";
@@ -228,7 +229,7 @@ void print_params(usrp_param my_parameter){
             ss << w_type_to_str(my_parameter.A_TXRX.wave_type[i]) <<std::endl;
         }
         ss << "   [#]     [linear]   [MHZ]      [MHz]      [#]     [msec]"<<std::endl<<std::endl;
-        
+
     }
     if(A_RX2){
         ss <<   "\033[40;1;32mRF A    RX2                                       No. of signals:    ";
@@ -238,12 +239,12 @@ void print_params(usrp_param my_parameter){
             ss <<                                   "      PFB points:   "<<my_parameter.A_RX2.fft_tones;
             ss <<                                                          "     PFB average:  "<<  my_parameter.A_RX2.pf_average<<" \033[0m"<<std::endl<<std::endl;
         }else{
-            ss <<"                                            "<<" \033[0m"<<std::endl<<std::endl;                     
+            ss <<"                                            "<<" \033[0m"<<std::endl<<std::endl;
         }
         ss << "\033[47;1;30m   No    \033[0m"<<" "<< "\033[47;1;30m   AMP   \033[0m"<<" "<<"\033[47;1;30m  FREQ   \033[0m"<<" ";
         ss << "\033[47;1;30m  2 FRQ  \033[0m"<<" "<< "\033[47;1;30m  STEPS  \033[0m"<<" "<<"\033[47;1;30m  LAPSE  \033[0m"<< "";
         ss << "\033[47;1;30m   TYPE  \033[0m"<<std::endl;
-        for(int i = 0; i<my_parameter.A_RX2.wave_type.size();i++){
+        for(size_t i = 0; i<my_parameter.A_RX2.wave_type.size();i++){
             ss << "    "<<(position_formatting % (i)).str() <<"      ";
             ss << (position_formatting % (my_parameter.A_RX2.ampl[i]/1.)).str() <<"   ";
             ss << (position_formatting % (my_parameter.A_RX2.freq[i]/1e6)).str() <<"   ";
@@ -258,7 +259,7 @@ void print_params(usrp_param my_parameter){
             ss << w_type_to_str(my_parameter.A_RX2.wave_type[i]) <<std::endl;
         }
         ss << "   [#]     [linear]   [MHZ]      [MHz]      [#]     [msec]"<<std::endl<<std::endl;
-        
+
     }
     if(B_TXRX){
         ss <<   "\033[40;1;32mRF B TX/RX                                        No. of signals:    ";
@@ -268,14 +269,14 @@ void print_params(usrp_param my_parameter){
             ss <<                                   "         PFB points:   "<<my_parameter.B_TXRX.fft_tones;
             ss <<                                                          "        PFB average:  "<<  my_parameter.B_TXRX.pf_average<<" \033[0m"<<std::endl<<std::endl;
         }else{
-            ss <<"                                            "<<" \033[0m"<<std::endl<<std::endl;                     
+            ss <<"                                            "<<" \033[0m"<<std::endl<<std::endl;
         }
         ss << "\033[47;1;30m   No    \033[0m"<<" "<< "\033[47;1;30m   AMP   \033[0m"<<" "<<"\033[47;1;30m  FREQ   \033[0m"<<" ";
         ss << "\033[47;1;30m  2 FRQ  \033[0m"<<" "<< "\033[47;1;30m  STEPS  \033[0m"<<" "<<"\033[47;1;30m  LAPSE  \033[0m"<< "";
         ss << "\033[47;1;30m   TYPE  \033[0m"<<std::endl;
-       
-        
-        for(int i = 0; i<my_parameter.B_TXRX.wave_type.size();i++){
+
+
+        for(size_t i = 0; i<my_parameter.B_TXRX.wave_type.size();i++){
             ss << "    "<<(position_formatting % (i)).str() <<"      ";
             ss << (position_formatting % (my_parameter.B_TXRX.ampl[i]/1.)).str() <<"   ";
             ss << (position_formatting % (my_parameter.B_TXRX.freq[i]/1e6)).str() <<"   ";
@@ -290,7 +291,7 @@ void print_params(usrp_param my_parameter){
             ss << w_type_to_str(my_parameter.B_TXRX.wave_type[i]) <<std::endl;
         }
         ss << "   [#]     [linear]   [MHZ]      [MHz]      [#]     [msec]"<<std::endl<<std::endl;
-        
+
     }
     if(B_RX2){
         ss <<   "\033[40;1;32mRF B    RX2                                       No. of signals:    ";
@@ -300,12 +301,12 @@ void print_params(usrp_param my_parameter){
             ss <<                                   "            PFB points:   "<<my_parameter.B_RX2.fft_tones;
             ss <<                                                             "        PFB average:  "<<  my_parameter.B_RX2.pf_average<<" \033[0m"<<std::endl<<std::endl;
         }else{
-            ss <<"                                            "<<" \033[0m"<<std::endl<<std::endl;                     
+            ss <<"                                            "<<" \033[0m"<<std::endl<<std::endl;
         }
         ss << "\033[47;1;30m   No    \033[0m"<<" "<< "\033[47;1;30m   AMP   \033[0m"<<" "<<"\033[47;1;30m  FREQ   \033[0m"<<" ";
         ss << "\033[47;1;30m  2 FRQ  \033[0m"<<" "<< "\033[47;1;30m  STEPS  \033[0m"<<" "<<"\033[47;1;30m  LAPSE  \033[0m"<< "";
         ss << "\033[47;1;30m   TYPE  \033[0m"<<std::endl;
-        for(int i = 0; i<my_parameter.B_RX2.wave_type.size();i++){
+        for(size_t i = 0; i<my_parameter.B_RX2.wave_type.size();i++){
             ss << "    "<<(position_formatting % (i)).str() <<"      ";
             ss << (position_formatting % (my_parameter.B_RX2.ampl[i]/1.)).str() <<"   ";
             ss << (position_formatting % (my_parameter.B_RX2.freq[i]/1e6)).str() <<"   ";
@@ -320,10 +321,10 @@ void print_params(usrp_param my_parameter){
             ss << w_type_to_str(my_parameter.B_RX2.wave_type[i]) <<std::endl;
         }
         ss << "   [#]     [linear]   [MHZ]      [MHz]      [#]     [msec]"<<std::endl<<std::endl;
-        
+
     }
     std::cout << ss.str();
-    
+
 }
 
 
@@ -430,4 +431,3 @@ void init_logger(){
     );
     */
 }
-
